Move alias class to separate package
