+++ binary-array-ld-lib/src/main/kotlin/net/bald/alias/AliasDefinition.kt
@@ -1,4 +1,4 @@
-package net.bald.context
+package net.bald.alias
 
 import org.apache.jena.rdf.model.Property
 import org.apache.jena.rdf.model.Resource
