+++ binary-array-ld-netcdf/src/main/kotlin/net/bald/netcdf/NetCdfContainer.kt
@@ -3,7 +3,7 @@ package net.bald.netcdf
 import net.bald.Attribute
 import net.bald.Container
 import net.bald.Var
-import net.bald.context.AliasDefinition
+import net.bald.alias.AliasDefinition
 import net.bald.context.ModelContext
 import org.apache.jena.rdf.model.Property
 import org.apache.jena.rdf.model.RDFNode
